test(consensus): share certificate fixtures
Moves DKG setup and certificate construction into shared test support so follower and tempo/1 tests exercise the same
fixtures.
